% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monoids.R
\name{mappend}
\alias{mappend}
\alias{\%<>\%}
\title{mappend}
\usage{
mappend(x, y)

x \%<>\% y
}
\arguments{
\item{x}{The left element}

\item{y}{The right element}
}
\value{
The result of appending
}
\description{
The monoidal append operation. Currently
implemented for:
\enumerate{
\item \code{list}s under concatenation,
\item \code{data.frame}s under row binding
\item \code{character} under \code{paste0}
\item \code{numeric} and \code{integer} under addition
}
}
\details{
appends two elements of the correct type.
}
