% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monoids.R
\name{mconcat}
\alias{mconcat}
\title{mconcat}
\usage{
mconcat(xs)
}
\arguments{
\item{xs}{A \code{list} or ocassionally \code{vector} of elements of
the type of the monoid.}
}
\value{
An element of the type of the monoid
}
\description{
Append a list (or occaisonally vector) of elements of
a monoid. Default implementation is to use \link[purrr:reduce]{purrr::reduce}
with \code{mappend} and \code{mempty} as the initial element. Fast
implementations are provided for each currently implemented
type.
}
