% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monoids.R
\name{mempty}
\alias{mempty}
\title{mempty}
\usage{
mempty(x)
}
\arguments{
\item{x}{An unused element indicating the desired type
of the identity element.}
}
\value{
An identity element of the correct type
}
\description{
The identity element of the monoid currently:
}
\details{
\enumerate{
\item \code{list()} for \code{list}s
\item \code{data.frame()} for \code{data.frame}s
\item \code{""} for \code{character}
\item \code{0} for \code{numeric} and \code{integer}
}
}
